Update update script to use sed to parse out latest version of EMBL file validator if XMLstarlet not installed. Use hard-coded version if failing.

+++ update_embl-validator.sh
@@ -12,22 +12,37 @@
 # missing from the current directory, or if it's outdated.
 
 # If XMLStarlet is installed, the script is smart and get the latest
-# version, otherwise it fetches the version corresponding to the value
-# of 'current_version' below.
+# version, otherwise it tries to use sed to parse the meta data
+# available.  If either of these approaches fail, the value of 'current_version' below is
+# used.
 
-current_version="1.1.146"
+current_version="1.1.147"
 
 if which -s xml; then
     # Use XMLStarlet to get the latest available version number.
-    current_version=$( curl http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/maven-metadata.xml | xml sel -t -v '//latest' -nl )
+    curr_version=$( curl -s http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/maven-metadata.xml | xml sel -t -v '//latest' -nl )
+
 else
-    # We don't have XMLStarlet installed, so use the hard-coded value.
+    # We don't have XMLStarlet installed, so hack it with sed instead.
+    curr_version=$( curl http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/maven-metadata.xml | sed -n '/latest/s/^[^0-9.]*\([0-9.]*\)[^0-9.]*$/\1/' )
 
-    # Please check the following URL for what the most current version is:
-    # http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/
-
-    true
 fi
 
-curl -z embl-validator.jar -o embl-validator.jar \
-    "http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/$current_version/embl-api-validator-$current_version.jar"
+if [ "x$curr_version" != "x" ]; then
+    current_version="$curr_version"
+    printf "ENA says current version of validator is '%s'\n" "$current_version"
+    echo "Fetching/updating it as needed..."
+else
+    echo "Failed in getting current version number of validator from ENA."
+    printf "Attempting to fetch version '%s'\n" "$current_version"
+fi
+
+if [ -f embl-validator.jar ]; then
+    # Updating existing validator JAR file.
+    curl -z embl-validator.jar -o embl-validator.jar \
+        "http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/$current_version/embl-api-validator-$current_version.jar"
+else
+    # Validator JAR file not present, fetching it.
+    curl -o embl-validator.jar \
+        "http://central.maven.org/maven2/uk/ac/ebi/ena/sequence/embl-api-validator/$current_version/embl-api-validator-$current_version.jar"
+fi
